Move GPG_TTY assignment to ~/.bashrc
When running under git, there is no controlling tty
which causes the 'tty' call inside git-remote-gcrypt
to fail.

Move the assignment to the shell initialization
so that it is set upon session startup. How this
interacts with having multiple shells, I don't
know yet, but it's enough for the trivial case.